[feat#1]장바구니기능추가

+++ src/main/java/cart/product/ProductRepository.java
@@ -3,6 +3,6 @@ package cart.product;
 import cart.product.model.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
 
 }
